+++ CHANGELOG.md
@@ -25,8 +25,8 @@
 * Pagination navigation item methods override now deprecated
 * Page switch methods returns a boolean to know if page switch was successful
 * Exception handler now catches page switch errors
-* Pagination `offset` and `limit` now deprecated
-* User-defined layout pattern can be applied per-context now
+* Paginated view `offset` and `limit` now deprecated
+* Fixed a bug that pagination data of multiple contexts in the same view conflicts with each other
 
 # 2.5
 * Inheritable context data (#65)